drcs reads from local, now to present
