+++ packages/volumetric/package.json
@@ -14,6 +14,7 @@
   "scripts": {},
   "author": "XR3 Group",
   "license": "MIT",
+  "type": "module",
   "bugs": {
     "url": "https://github.com/xr3ngine/three-volumetric/issues"
   },
@@ -27,7 +28,9 @@
   "dependencies": {
     "cross-blob": "^2.0.0",
     "fs-read-stream-over-http": "^0.2.2",
-    "xmlhttprequest": "^1.8.0",
-    "three": "^0.123.0"
+    "http-stream": "^0.1.1",
+    "stream-http": "^3.1.1",
+    "three": "^0.123.0",
+    "xmlhttprequest": "^1.8.0"
   }
 }
